common: use automatic address family detection for server sockets
Rather than creating an AF_INET socket, create a socket with the
most-preferred operating system AF.

_Only the first entry_ of getaddrinfo's result is inspected for now.
This means that if IPv6 is available in the system, it will be
chosen for the socket type, and no IPv4 socket will be
created. However, since IPv6 sockets by default also handle v4 traffic,
this should reasonably work.

References: ZCP-6702
Originally from SVN commits 43554,43556,43600

+++ common/include/zarafa/ECChannel.h
@@ -102,7 +102,7 @@ private:
 
 /* helpers to open socket */
 HRESULT HrListen(ECLogger *lpLogger, const char *szPath, int *lpulListenSocket);
-HRESULT HrListen(ECLogger *lpLogger, const char *szBind, int ulPort, int *lpulListenSocket);
+HRESULT HrListen(ECLogger *lpLogger, const char *szBind, uint16_t ulPort, int *lpulListenSocket);
 /* accept data on connection */
 HRESULT HrAccept(ECLogger *lpLogger, int ulListenFD, ECChannel **lppChannel);
